Split search controller
