+++ api/catalog/api/views/media_views.py
@@ -12,6 +12,7 @@ from rest_framework.response import Response
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
 from catalog.api.controllers import search_controller
+from catalog.api.controllers.elasticsearch.stats import get_stats
 from catalog.api.models import ContentProvider
 from catalog.api.serializers.provider_serializers import ProviderSerializer
 from catalog.api.utils.exceptions import get_api_exception
@@ -80,7 +81,7 @@ class MediaViewSet(ReadOnlyModelViewSet):
 
     @action(detail=False, serializer_class=ProviderSerializer, pagination_class=None)
     def stats(self, *_, **__):
-        source_counts = search_controller.get_sources(self.default_index)
+        source_counts = get_stats(self.default_index)
         context = self.get_serializer_context() | {
             "source_counts": source_counts,
         }
